fix: organize maps components in own folder
